release

+++ Source/UniqueEngine/Export/GraphicsExport.cpp
@@ -37,7 +37,7 @@ UNIQUE_C_API void StaticModel_SetMaterialsAttr(StaticModel* self, ResourceRefLis
 	self->SetMaterialsAttr(*material);
 }
 
-uExport(Graphics, float, GetWidth)
+uExport(Graphics, int, GetWidth)
 uExport(Graphics, int, GetHeight)
 
 uExportStatic(Graphics, float, GetRenderWait)
